Add 'receive' feature
It can be turned off for devices that only send, e.g. online controls
placed in the forest.
