+++ Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.13"
 edition = "2021"
 rust-version = "1.70"
 
+[features]
+default = ["receive"]
+receive = ["dep:meshtastic"]
+
 [build-dependencies]
 prost-build = "=0.13.2"
 protoc-bin-vendored = "3.1.0"
@@ -13,7 +17,7 @@ walkdir = "2.3.2"
 chrono = "0.4.38"
 geoutils = "0.5.1"
 log = "0.4.20"
-meshtastic = { git = "https://github.com/meshtastic/rust.git", rev = "7410cefb3747ee7c8b6dc81a182ed91b810f9e20" }
+meshtastic = { git = "https://github.com/meshtastic/rust.git", rev = "7410cefb3747ee7c8b6dc81a182ed91b810f9e20", optional = true }
 prost = "=0.13.2"
 pyo3 = { version = "0.22.0", features = ["extension-module", "chrono", "abi3-py311"] }
 pyo3-log = "0.11.0"
