+++ src/lib.rs
@@ -1,5 +1,8 @@
 //pub mod async_serial;
 pub mod logs;
+#[cfg(feature = "receive")]
+pub mod meshtastic;
+#[cfg(feature = "receive")]
 pub mod message_handler;
 pub mod punch;
 pub mod python;
